prepare_db_for_init の単体テストを test_run_unit.py に移動する
